feat: move frontend types to shared types repo
