+++ apps/web/types/index.d.ts
@@ -1,18 +1,6 @@
-import { Dispatch, SetStateAction } from 'react'
-
-export type SetState<Type> = Dispatch<SetStateAction<Type>>
-export type UseState<Type> = [Type, SetState<Type>]
-
-export type FlowNodeCondensed = {
-  moduleCode: string
-  position: {
-    x: number
-    y: number
-  }
-}
-
-export type FlowEdgeCondensed = {
-  id: string
-  source: string
-  target: string
-}
+/**
+ * for frontend-specific types.
+ * importable from within this project with
+ * `import { Placeholder } from 'types'`
+ */
+export type Placeholder = string
