+++ libs/types/src/flow.ts
@@ -1,0 +1,13 @@
+export type FlowNodeCondensed = {
+  moduleCode: string
+  position: {
+    x: number
+    y: number
+  }
+}
+
+export type FlowEdgeCondensed = {
+  id: string
+  source: string
+  target: string
+}
