+++ libs/types/src/react.ts
@@ -1,0 +1,4 @@
+import { Dispatch, SetStateAction } from 'react'
+
+export type SetState<Type> = Dispatch<SetStateAction<Type>>
+export type UseState<Type> = [Type, SetState<Type>]
